informal-settlements-Amazon
The idea of this model is to explore the informal settlers’ migration preference, building on the understanding of the local reality of informal settlers from a field trip in Brazilian Amazon. GIS vector data representing real world geographic space was imported in NetLogo. Initializing the informal settlements is based on the empirical evidence that most of the informal settlers choose the nearest water resources; primary migration considers a proportion of people’s intention of moving closer to the cities nearby; and secondary migration enables people’s ability to evaluate the environmental conditions (represented by adjustable factors: water, transportation, population, and land value), and make further migration decisions. This model makes a good attempt to visualize geospatial realities, but still simplifies the factors. Nevertheless, it gives a sense of how people will adapted to their living conditions, and will be aggregated spatially. There is still a huge potential to simulate much more complex realities with more detailed information and more coding effort.